Moving expiration tests to its own test file
